+++ changelog.md
@@ -6,6 +6,7 @@
 
 * Fixed `lk.download_from_doi()` to align with new Zenodo REST API.
 * Fixed a bug in `Scan.plot()` in which the default aspect ratio was calculated such that pixels always appeared square. For scans with non-square pixel sizes, this would result in distortion of the image.
+* Don't store animation writer in a temporary variable as this results in a `matplotlib` error when attempting to export a movie on jupyter notebook. 
 
 ## v1.2.1 | 2023-10-17